+++ lib/atom-items/issueish-tooltip-item.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDom from 'react-dom';
-import {QueryRenderer, graphql} from 'react-relay/compat';
+import {QueryRenderer, graphql} from 'react-relay';
 
 import IssueishTooltipContainer from '../containers/issueish-tooltip-container';
