+++ lib/atom-items/user-mention-tooltip-item.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDom from 'react-dom';
-import {QueryRenderer, graphql} from 'react-relay/compat';
+import {QueryRenderer, graphql} from 'react-relay';
 
 import UserMentionTooltipContainer from '../containers/user-mention-tooltip-container';
