+++ lib/containers/issueish-pane-item-container.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {graphql, createRefetchContainer} from 'react-relay/compat';
+import {graphql, createRefetchContainer} from 'react-relay';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import {autobind} from 'core-decorators';
