+++ lib/containers/pr-status-context-container.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {createFragmentContainer, graphql} from 'react-relay/compat';
+import {createFragmentContainer, graphql} from 'react-relay';
 import PropTypes from 'prop-types';
 
 import Octicon from '../views/octicon';
