+++ lib/containers/pr-timeline-container.js
@@ -1,4 +1,4 @@
-import {graphql, createPaginationContainer} from 'react-relay/compat';
+import {graphql, createPaginationContainer} from 'react-relay';
 
 import IssueishTimelineView from '../views/issueish-timeline-view';
 import CommitsContainer from './../containers/timeline-items/commits-container.js';
