+++ lib/containers/timeline-items/commit-container.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {graphql, createFragmentContainer} from 'react-relay/compat';
+import {graphql, createFragmentContainer} from 'react-relay';
 import PropTypes from 'prop-types';
 
 import Octicon from '../../views/octicon';
