+++ lib/controllers/issueish-pane-item-controller.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import yubikiri from 'yubikiri';
 import {autobind} from 'core-decorators';
-import {QueryRenderer, graphql} from 'react-relay/compat';
+import {QueryRenderer, graphql} from 'react-relay';
 
 import RelayNetworkLayerManager from '../relay-network-layer-manager';
 import GithubLoginModel, {UNAUTHENTICATED} from '../models/github-login-model';
