+++ lib/controllers/pr-info-controller.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {QueryRenderer, graphql} from 'react-relay/compat';
+import {QueryRenderer, graphql} from 'react-relay';
 import {autobind} from 'core-decorators';
 
 import {RemotePropType} from '../prop-types';
