+++ lib/relay-network-layer-manager.js
@@ -1,33 +1,24 @@
-import Relay from 'react-relay/classic';
+import {Environment, Network, RecordSource, Store} from 'relay-runtime';
 
 const relayEnvironmentPerGithubHost = new Map();
 
-function override(target, method, implementation) {
-  const oldImplementation = target[method].bind(target);
-  target[method] = implementation(oldImplementation);
-}
-
-function sendQueriesOverride(originalSendQueries) {
-  return function sendQueriesWithRateLimitReq(requests) {
-    requests.forEach(request => {
-      override(request, 'getQueryString', getQueryStringOverride);
-      request.then(data => logRateLimitInfo(data.response));
+function createFetchQuery(url, token) {
+  return function fetchQuery(operation, variables, cacheConfig, uploadables) {
+    return fetch(url, {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+        'Authorization': `bearer ${token}`,
+        'Accept': 'application/vnd.github.graphql-profiling+json',
+      },
+      body: JSON.stringify({
+        query: operation.text,
+        variables,
+      }),
+    }).then(response => {
+      return response.json();
     });
-    return originalSendQueries(requests);
   };
-}
-
-function getQueryStringOverride(originalGetQueryString) {
-  return function getQueryStringInjectingRateLimitReq() {
-    const str = originalGetQueryString();
-    return str.replace(/query (.*\{)/, 'query $1 ghRateLimit:relay { rateLimit { limit cost remaining resetAt } } ');
-  };
-}
-
-function logRateLimitInfo(response) {
-  if (response && response.ghRateLimit) {
-    console.table(response.ghRateLimit); // eslint-disable-line no-console
-  }
 }
 
 export default class RelayNetworkLayerManager {
@@ -35,24 +26,15 @@ export default class RelayNetworkLayerManager {
     host = host === 'github.com' ? 'https://api.github.com' : host; // eslint-disable-line no-param-reassign
     const url = host === 'https://api.github.com' ? `${host}/graphql` : `${host}/api/v3/graphql`;
     const config = relayEnvironmentPerGithubHost.get(host) || {};
-    let {environment, networkLayer} = config;
+    let {environment, network} = config;
     if (!environment) {
-      environment = new Relay.Environment();
-      networkLayer = new Relay.DefaultNetworkLayer(url, {
-        headers: {
-          Authorization: `bearer ${token}`,
-          Accept: 'application/vnd.github.graphql-profiling+json',
-        },
-      });
-      if (atom && atom.inDevMode()) {
-        override(networkLayer, 'sendQueries', sendQueriesOverride);
-      }
-      environment.injectNetworkLayer(networkLayer);
-      relayEnvironmentPerGithubHost.set(host, {environment, networkLayer});
+      const source = new RecordSource();
+      const store = new Store(source);
+      network = Network.create(createFetchQuery(url, token));
+      environment = new Environment({network, store});
+
+      relayEnvironmentPerGithubHost.set(host, {environment, network});
     } else {
-      networkLayer._init.headers = {
-        Authorization: `bearer ${token}`,
-      };
     }
     return environment;
   }
